Fix crash when resizing window in map editor
Something is processing events during render, but finding it is
more difficult than just aborting drawing when the vector of
drawables gets invalidated.

Once that is fixed nothing will be changing the drawable vector,
so this will then have no effect. Until then it will cause some
flickering when resizing, but it won't crash.